LayoutTests/js/code-serialize-paren-expected.txt

This test checks whether converting function code to a string preserves semantically significant parentheses.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS (function () { return (x + y) * z; }).toString().search('return.*[(]') != -1 is true
PASS successfullyParsed is true

TEST COMPLETE
